refactor(store): 매장 검색 API, url mapping 변경
- api/customer/store 로 url mapping 변경
- 테스트 url 변경
